refactor(cli): inline fsx into cli package
Drop the separate cli/fsx subpackage: WriteFS and OS() now live in
cli/fs.go. cli/repl declares its own WriteFS interface on the consumer
side (the only place that uses it) so the REPL package stays cycle-free
when cli imports it. The REPL no longer carries a default filesystem;
cli.Root injects cli.OS() at the entry point, and standalone repl.New
disables .load/.save unless WithFS is supplied.
